package refactor to better accomodate better auth provider
