+++ packages/cms-core/src/api/index.ts
@@ -1,4 +1,4 @@
 // API Client exports
 export { ApiClient, ApiError, apiClient } from './client.js';
 export { documents } from './documents.js';
-export type * from './types.js';
+export type * from './types/index.js';
