+++ packages/cms-core/src/app.d.ts
@@ -1,8 +1,8 @@
 // See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
 //
-import type { CMSInstances } from '@aphex/cms-core/server';
-import type { Auth } from 'better-auth';
+import type { CMSInstances } from './hooks.js';
+import type { Auth } from './types/index.js';
 declare global {
     namespace App {
         // interface Error {}
