+++ packages/cms-core/src/db/adapters/index.ts
@@ -1,3 +1,3 @@
 // Database adapters are now in separate packages
 // e.g., @aphex/postgresql-adapter, @aphex/mongodb-adapter
-// This file is kept for backwards compatibility but exports nothing
+// This file is kept for backwards compatibility but exports nothing - and potentially default adapters or something. maybe sqlite?
