+++ packages/cms-core/src/define.ts
@@ -1,5 +1,5 @@
 // CMS configuration helper
-import type { CMSConfig } from './types.js';
+import type { CMSConfig } from './types/index.js';
 
 // CMS config helper
 export function defineCMSConfig(config: CMSConfig): CMSConfig {
