--- packages/cms-core/src/services/factory.ts
@@ -1,8 +1,0 @@
-// Service factory based on configuration
-import type { CMSConfig } from '../config.js';
-
-export function createServices(config: CMSConfig) {
-	// This will be implemented when we need it
-	// For now, services are created in the hooks
-	return {};
-}
