+++ packages/cms-core/src/types/schemas.ts
@@ -1,0 +1,137 @@
+// types/schemas.ts
+import type { Rule } from '../field-validation/rule.js';
+
+// From root types.ts
+export type FieldType =
+	| 'string'
+	| 'text'
+	| 'number'
+	| 'boolean'
+	| 'slug'
+	| 'image'
+	| 'array'
+	| 'object'
+	| 'reference';
+
+export interface BaseField {
+	name: string;
+	type: FieldType;
+	title: string;
+	description?: string;
+	validation?: ((rule: Rule) => Rule) | Array<(rule: Rule) => Rule>;
+}
+
+export interface StringField extends BaseField {
+	type: 'string';
+	maxLength?: number;
+	placeholder?: string;
+}
+
+export interface TextField extends BaseField {
+	type: 'text';
+	rows?: number;
+	maxLength?: number;
+	placeholder?: string;
+}
+
+export interface NumberField extends BaseField {
+	type: 'number';
+	min?: number;
+	max?: number;
+}
+
+export interface BooleanField extends BaseField {
+	type: 'boolean';
+	initialValue?: boolean;
+}
+
+export interface SlugField extends BaseField {
+	type: 'slug';
+	source?: string;
+	maxLength?: number;
+}
+
+export interface ImageField extends BaseField {
+	type: 'image';
+	accept?: string;
+	hotspot?: boolean; // Enable hotspot/crop UI
+	metadata?: string[]; // e.g., ['palette', 'exif', 'location']
+	fields?: Field[]; // Additional fields like caption, attribution
+}
+
+export interface TypeReference {
+	type: string; // References a SchemaType by name
+	title?: string;
+}
+
+export interface ArrayField extends BaseField {
+	type: 'array';
+	of: TypeReference[];
+	max?: number;
+}
+
+export interface ObjectField extends BaseField {
+	type: 'object';
+	fields: Field[];
+}
+
+export interface ReferenceField extends BaseField {
+	type: 'reference';
+	to: Array<{ type: string }>;
+}
+
+export type Field =
+	| StringField
+	| TextField
+	| NumberField
+	| BooleanField
+	| SlugField
+	| ImageField
+	| ArrayField
+	| ObjectField
+	| ReferenceField;
+
+export interface DocumentType {
+	type: 'document';
+	name: string;
+	title: string;
+	description?: string;
+	fields: Field[];
+}
+
+export interface ObjectType {
+	type: 'object';
+	name: string;
+	title: string;
+	description?: string;
+	fields: Field[];
+}
+
+// From db/types.ts
+/**
+ * Schema type - represents document and object type definitions stored in the DB
+ */
+export interface SchemaType {
+	id: string;
+	name: string;
+	title: string;
+	type: 'document' | 'object';
+	description: string | null;
+	schema: any;
+	createdAt: Date | null;
+	updatedAt: Date | null;
+}
+
+/**
+ * New schema type input type
+ */
+export interface NewSchemaType {
+	id?: string;
+	name: string;
+	title: string;
+	type: 'document' | 'object';
+	description?: string | null;
+	schema: any;
+	createdAt?: Date | null;
+	updatedAt?: Date | null;
+}
